Support --root for `npm start`
This overlays the --root directories over the reveal.js source directory,
allowing simple sharing of reveal.js amongst multiple presentations.

This appears to be a bug-fix, as this was already attempting to serve
the root directories, but when doing so, the reveal.js directories were
not visible.

Either that, or the --root parameter was only expected to be used when
reveal.js was within the hiearchy and accessed using a relative path
from the index.html